refactor(fetch-llm-completion): consolidate OpenAI AI SDK code into providers module
Move buildOpenAIModel and the provider-options translation into
providers/openai.ts so the directory structure marks where future AI SDK
adapters (e.g. Anthropic) plug in. Pure code motion; no behavior change.
